Refactor - add right click menu component
